Update Mesh to use ValueVector.
Modify ElementFem ctor to accept nodes by reference_wrapper. This avoids
the pointer syntax (the & sign in ElementFem({&n0...}).
