+++ src/mechanics/elements/ElementFem.h
@@ -12,12 +12,13 @@ namespace NuTo
 class ElementFem : public ElementInterface
 {
 public:
-    ElementFem(std::vector<NuTo::NodeSimple*> nodes, const InterpolationSimple& interpolation)
-        : mNodes(nodes)
-        , mInterpolation(interpolation)
+    ElementFem(std::initializer_list<std::reference_wrapper<NuTo::NodeSimple>> nodes, const InterpolationSimple& interpolation)
+        : mInterpolation(interpolation)
     {
-        assert(nodes.size() == interpolation.GetNumNodes());
-        assert(nodes.front()->GetNumValues() == interpolation.GetDofDimension());
+        for (NuTo::NodeSimple& node : nodes)
+            mNodes.push_back(&node);
+        assert(mNodes.size() == interpolation.GetNumNodes());
+        assert(mNodes.front()->GetNumValues() == interpolation.GetDofDimension());
     }
 
     virtual NodeValues ExtractNodeValues() const override
