+++ test/mechanics/cell/Cell.cpp
@@ -27,14 +27,14 @@ BOOST_AUTO_TEST_CASE(CellLetsSee)
     NuTo::NodeSimple nCoord1(Eigen::Vector2d({lx, 0}));
     NuTo::NodeSimple nCoord2(Eigen::Vector2d({lx, ly}));
     NuTo::NodeSimple nCoord3(Eigen::Vector2d({0, ly}));
-    NuTo::ElementFem coordinateElement({&nCoord0, &nCoord1, &nCoord2, &nCoord3}, interpolationCoordinates);
+    NuTo::ElementFem coordinateElement({nCoord0, nCoord1, nCoord2, nCoord3}, interpolationCoordinates);
 
     NuTo::InterpolationQuadLinear interpolationDisplacements(2);
     NuTo::NodeSimple nDispl0(Eigen::Vector2d({0, 0}));
     NuTo::NodeSimple nDispl1(Eigen::Vector2d({0, 0}));
     NuTo::NodeSimple nDispl2(Eigen::Vector2d({0, 0}));
     NuTo::NodeSimple nDispl3(Eigen::Vector2d({0, 0}));
-    NuTo::ElementFem displacementElement({&nDispl0, &nDispl1, &nDispl2, &nDispl3}, interpolationDisplacements);
+    NuTo::ElementFem displacementElement({nDispl0, nDispl1, nDispl2, nDispl3}, interpolationDisplacements);
 
     NuTo::ElementCollectionFem elements(coordinateElement);
     NuTo::DofType dofDispl("Displacements", 2);
