+++ test/mechanics/elements/ElementCollection.cpp
@@ -11,7 +11,7 @@ BOOST_AUTO_TEST_CASE(ElementCollectionAccess)
     NuTo::InterpolationTriangleLinear interpolation(2);
 
 
-    NuTo::ElementCollectionFem elements(NuTo::ElementFem({&n0, &n1, &n2}, interpolation));
+    NuTo::ElementCollectionFem elements(NuTo::ElementFem({n0, n1, n2}, interpolation));
     BOOST_CHECK_NO_THROW(elements.CoordinateElement().Interpolation());
 
     NuTo::ElementCollection& cellElements = elements;
