+++ test/mechanics/mesh/MeshFem.cpp
@@ -1,0 +1,29 @@
+#include "BoostUnitTest.h"
+#include "mechanics/mesh/MeshFem.h"
+#include "mechanics/interpolation/InterpolationTriangleLinear.h"
+
+BOOST_AUTO_TEST_CASE(MeshAddStuff)
+{
+    NuTo::MeshFem mesh;
+    NuTo::InterpolationTriangleLinear interpolationTriangle(2);
+    auto& interpolation = mesh.CreateInterpolation(interpolationTriangle);
+
+    BOOST_CHECK_EQUAL(interpolation.GetNumNodes(), 3);
+    BOOST_CHECK_EQUAL(interpolation.GetDofDimension(), 2);
+
+
+    auto& n0 = mesh.Nodes.Add(Eigen::Vector2d({1, 0}));
+    auto& n1 = mesh.Nodes.Add(Eigen::Vector2d({2, 0}));
+    auto& n2 = mesh.Nodes.Add(Eigen::Vector2d({0, 3}));
+
+    BOOST_CHECK_EQUAL(n0.GetNumValues(), 2);
+    BOOST_CHECK_EQUAL(n1.GetNumValues(), 2);
+    BOOST_CHECK_EQUAL(n2.GetNumValues(), 2);
+
+    auto& e0 = mesh.Elements.Add({{{n0, n1, n2}, interpolation}}); 
+    // The slightly wierd syntax, double {{}} is caused by the forwarding Add.
+    // Alternatively:
+    // auto& e0 = mesh.Elements.Add(NuTo::ElementFem({n0, n1, n2}, interpolation)); 
+
+    BoostUnitTest::CheckVector(e0.CoordinateElement().ExtractNodeValues(), std::vector<double>({1, 0, 2, 0, 0, 3}), 6);
+}
